Move non-Cohort objects out of load.py
